moved discord link flow to url handler
