+++ frontend/src/ts/utils/url-handler.ts
@@ -5,7 +5,51 @@ import { decompressFromURI } from "lz-ts";
 import * as QuoteSearchPopup from "../popups/quote-search-popup";
 import * as ManualRestart from "../test/manual-restart-tracker";
 import * as CustomText from "../test/custom-text";
+import Ape from "../ape";
+import * as Settings from "../pages/settings";
+import * as DB from "../db";
+import * as Loader from "../elements/loader";
+import * as AccountButton from "../elements/account-button";
 import { restart as restartTest } from "../test/test-logic";
+
+export async function linkDiscord(hashOverride: string): Promise<void> {
+  if (!hashOverride) return;
+  const fragment = new URLSearchParams(hashOverride.slice(1));
+  if (fragment.has("access_token")) {
+    history.replaceState(null, "", "/");
+    const accessToken = fragment.get("access_token") as string;
+    const tokenType = fragment.get("token_type") as string;
+
+    Loader.show();
+
+    const response = await Ape.users.linkDiscord(tokenType, accessToken);
+    Loader.hide();
+
+    if (response.status !== 200) {
+      return Notifications.add(
+        "Failed to link Discord: " + response.message,
+        -1
+      );
+    }
+
+    Notifications.add(response.message, 1);
+
+    const snapshot = DB.getSnapshot();
+
+    const { discordId, discordAvatar } = response.data;
+    if (discordId) {
+      snapshot.discordId = discordId;
+    } else {
+      snapshot.discordAvatar = discordAvatar;
+    }
+
+    DB.setSnapshot(snapshot);
+
+    AccountButton.update(discordId, discordAvatar);
+
+    Settings.updateDiscordSection();
+  }
+}
 
 export function loadCustomThemeFromUrl(getOverride?: string): void {
   const getValue = Misc.findGetParameter("customTheme", getOverride);
